Use dynamic annotated fields in the clipboard.
